% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsiex.termdoc.data.R
\docType{data}
\name{lsiex.termdoc}
\alias{lsiex.termdoc}
\title{Document terms from lsiex data set}
\format{A data frame of 384x17}
\description{
Used in Example 2-3 for Singular Value Decomposition
}
\examples{
lsiex.termdoc[1]
}
\keyword{datasets}
