% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch01.R
\name{m.eig}
\alias{m.eig}
\title{EigenValues Function}
\usage{
m.eig(M)
}
\arguments{
\item{M}{a matrix}
}
\description{
This function replicate the Matlab eig() function.
}
\details{
Returns:
      the diagonal of eigen$values.
      the negative of eigen$vectors in descending order.
}
\examples{

}
\keyword{EDA}
