% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{m.rand}
\alias{m.rand}
\title{Replicates Matlab's rand(m, n)}
\usage{
m.rand(m = 1, n = 1)
}
\arguments{
\item{m}{number of rows}

\item{n}{number of columns}
}
\description{
Matrix of mxn uniform distribution numbers
}

