% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch02.R
\name{m.svd}
\alias{m.svd}
\title{Replicate Matlab function svd including the diag(d) which is not
    included in R svd()}
\usage{
m.svd(M)
}
\arguments{
\item{M}{a matrix}
}
\description{
Replicate Matlab function svd including the diag(d) which is not
    included in R svd()
}
