% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch02-plot.R
\name{projectionMatrixLine}
\alias{projectionMatrixLine}
\title{Draw Projection Matrix 2D Line}
\usage{
projectionMatrixLine(M, ...)
}
\arguments{
\item{M}{projection matrix}

\item{...}{other parameters for \code{\link{abline}}}
}
\description{
draw the projection line given the projection matrix of angle theta
}

