% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.mvnrnd.R
\docType{data}
\name{X.mvnrnd}
\alias{X.mvnrnd}
\alias{x}
\title{Random number for the normal distribution.}
\format{A data frame 100x2 of random numbers}
\description{
Random numbers coming from Matlab example.
}
\details{
\itemize{
\item \code{V1} - column 1
\item \code{V2} - column 2
}
}
\examples{
X.mvnrnd[5]
}
\keyword{datasets}
