% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.data.R
\docType{data}
\name{yeast.data}
\alias{yeast.data}
\title{Random number for the normal distribution.}
\format{A data frame of 384x17}
\description{
Random numbers coming from Matlab example.
}
\examples{
yeast.data[1]
}
\keyword{datasets}
